+++ specification/ml/_types/Analysis.ts
@@ -17,11 +17,13 @@
  * under the License.
  */
 
-import { CategorizationAnalyzer } from '@ml/info/CategorizationAnalyzer'
 import { Field } from '@_types/common'
 import { long } from '@_types/Numeric'
 import { Time, TimeSpan } from '@_types/Time'
 import { Detector } from './Detector'
+import { CharFilter } from '@_types/analysis/char_filters'
+import { Tokenizer } from '@_types/analysis/tokenizers'
+import { TokenFilter } from '@_types/analysis/token_filters'
 
 export class AnalysisConfig {
   bucket_span: TimeSpan
@@ -57,4 +59,10 @@ export class AnalysisMemoryLimit {
    * Limits can be applied for the resources required to hold the mathematical models in memory. These limits are approximate and can be set per job. They do not control the memory used by other processes, for example the Elasticsearch Java processes.
    */
   model_memory_limit: string
+}
+
+export class CategorizationAnalyzer {
+  filter?: Array<string | TokenFilter>
+  tokenizer?: string | Tokenizer
+  char_filter?: Array<string | CharFilter>
 }
