+++ specification/ml/estimate_model_memory/MlEstimateModelMemoryResponse.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-import { ByteSize } from "@_types/common";
+import { ByteSize } from '@_types/common'
 
 export class Response {
   body: { model_memory_estimate: ByteSize }
