+++ specification/ml/info/MlInfoRequest.ts
@@ -24,6 +24,4 @@ import { RequestBase } from '@_types/Base'
  * @since 6.3.0
  * @stability TODO
  */
-export interface Request extends RequestBase {
-  query_parameters?: {}
-}
+export interface Request extends RequestBase {}
